reset table to page 1;
