+++ ui/app/components/clients/table.ts
@@ -5,7 +5,7 @@
 
 import Component from '@glimmer/component';
 import { action } from '@ember/object';
-import { tracked } from '@glimmer/tracking';
+import { cached, tracked } from '@glimmer/tracking';
 import { paginate } from 'core/utils/paginate-list';
 
 /**
@@ -50,22 +50,32 @@ export default class ClientsTable extends Component<Args> {
   @tracked currentPage = 1;
   @tracked pageSize = 5; // Can be overridden by @setPageSize
   @tracked sortColumn = '';
-  @tracked sortDirection: SortDirection;
+  @tracked sortDirection: SortDirection = 'asc'; // default is 'asc' for consistency with HDS defaults
 
   constructor(owner: unknown, args: Args) {
     super(owner, args);
 
-    const { column = '', direction = 'asc' } = this.args.initiallySortBy || {};
-    this.sortColumn = column;
-    this.sortDirection = direction; // default is 'asc' for consistency with HDS defaults
+    const { column, direction } = this.args.initiallySortBy || {};
+    if (column) {
+      this.sortColumn = column;
+    }
+    if (direction) {
+      this.sortDirection = direction;
+    }
 
     // Override default page size with a custom amount.
-    // pageSize can be updated by the end user if @showPaginationSizeSelector is true
+    // pageSize can be updated by the user if @showPaginationSizeSelector is true
     if (this.args.setPageSize) {
       this.pageSize = this.args.setPageSize;
     }
   }
 
+  @cached
+  get columnKeys() {
+    return this.args.columns.map((k: TableColumn) => k['key']);
+  }
+
+  @cached
   get paginatedTableData(): Record<string, any>[] {
     const sorted = this.sortTableData(this.args.data);
     const paginated = paginate(sorted, {
@@ -74,10 +84,6 @@ export default class ClientsTable extends Component<Args> {
     });
 
     return paginated;
-  }
-
-  get columnKeys() {
-    return this.args.columns.map((k: TableColumn) => k['key']);
   }
 
   // This table is paginated so we cannot use any out of the box filtering
@@ -99,6 +105,13 @@ export default class ClientsTable extends Component<Args> {
   @action
   handlePaginationChange(action: 'currentPage' | 'pageSize', value: number) {
     this[action] = value;
+  }
+
+  @action
+  resetPagination() {
+    // setPageSize is intentionally NOT reset here so user changes to page size
+    // are preserved regardless of whether or not the table data updates.
+    this.currentPage = 1;
   }
 
   @action
